requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
alabaster==0.7.12         # via sphinx
argon2-cffi==20.1.0       # via keyrings.cryptfile
babel==2.8.0              # via sphinx
certifi==2020.6.20        # via requests
cffi==1.14.3              # via argon2-cffi, cryptography
chardet==3.0.4            # via requests
cryptography==3.2.1       # via secretstorage
docutils==0.16            # via sphinx
idna==2.10                # via requests
imagesize==1.2.0          # via sphinx
imutils==0.5.3            # via -r requirements.in
jeepney==0.4.3            # via keyring, secretstorage
jinja2==2.11.2            # via sphinx
keyring==21.5.0           # via keyrings.cryptfile
keyrings.cryptfile==1.3.6  # via -r requirements.in
markupsafe==1.1.1         # via jinja2
nose==1.3.7               # via -r requirements.in
numpy==1.19.4             # via opencv-python
opencv-python==4.4.0.46   # via -r requirements.in
packaging==20.4           # via sphinx
pillow==8.0.1             # via -r requirements.in
pycparser==2.20           # via cffi
pycryptodome==3.9.9       # via keyrings.cryptfile
pygments==2.7.2           # via sphinx
pyparsing==2.4.7          # via packaging
python-levenshtein-wheels==0.13.1  # via -r requirements.in
python-slugify==4.0.1     # via -r requirements.in
pytz==2020.4              # via babel
requests==2.24.0          # via -r requirements.in, sphinx
secretstorage==3.2.0      # via keyring
six==1.15.0               # via argon2-cffi, cryptography, packaging
snowballstemmer==2.0.0    # via sphinx
sphinx==3.3.0             # via -r requirements.in
sphinxcontrib-applehelp==1.0.2  # via sphinx
sphinxcontrib-devhelp==1.0.2  # via sphinx
sphinxcontrib-htmlhelp==1.0.3  # via sphinx
sphinxcontrib-jsmath==1.0.1  # via sphinx
sphinxcontrib-qthelp==1.0.3  # via sphinx
sphinxcontrib-serializinghtml==1.1.4  # via sphinx
tesserocr==2.5.1          # via -r requirements.in
text-unidecode==1.3       # via python-slugify
urllib3==1.25.11          # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
